fix bug checkout (#26)
* Dev 2 (#22)

* Dev 1 (#21)

* config inertia

* config tailwind and laravel sptie

* config permissions

* create migration and seeder

* fix primary key

* change id to uuid

* add register, login, logout (#4)

* create  profile account

* fix table id because permission issue

* create layout account

* create profile account

* fix some issue

* roles and permissions

* modify stling and props

* create permissions feature

* create roles feature

* create manage users

* stores delivery 

* create manage stores

* create manage delivery areas

* create manage sliders and categories

* create manage product and order

* create dashboard

* create home, product,category page

* update git ignore

* update vite config.js

* create cart and ceckout

* Dev 2.2 (#24) (#25)

* Dev 2 (#22)

* Dev 1 (#21)

* config inertia

* config tailwind and laravel sptie

* config permissions

* create migration and seeder

* fix primary key

* change id to uuid

* add register, login, logout (#4)

* create  profile account

* fix table id because permission issue

* create layout account

* create profile account

* fix some issue

* roles and permissions

* modify stling and props

* create permissions feature

* create roles feature

* create manage users

* stores delivery 

* create manage stores

* create manage delivery areas

* create manage sliders and categories

* create manage product and order

* create dashboard

* create home, product,category page

* update git ignore

* update vite config.js

* create cart and ceckout

* fix checkout function

+++ app/Http/Middleware/HandleInertiaRequests.php
@@ -45,7 +45,7 @@ class HandleInertiaRequests extends Middleware
             //carts
             'dataCarts' => $request->user() ? [
                 'total'     =>  \App\Models\Cart::where('user_id', $request->user()->id)->count() ?? 0,
-                'price'     => (float) \App\Models\Cart::where('user_id', $request->user()->id)->sum('total_price') ?? 0,
+                'price'     => (int) \App\Models\Cart::where('user_id', $request->user()->id)->sum('total_price') ?? 0,
             ] : null
         ]);
     }
